break up FfprobeVersion into Ffprober::Ffmpeg::VersionValidator and Ffprober::Ffmpeg::Version
