+++ spec/ffprober/ffmpeg/finder_spec.rb
@@ -1,4 +1,3 @@
-# encoding: utf-8
 require 'spec_helper'
 
 describe Ffprober::Ffmpeg::Finder do
